Add functionality to include data with http request
Data can be included in the http request as the request body. Currently
only json is supported, all other data is send as plane text. One should
specify the content type explictly, by setting Content-Type header. If
content-type is not specified, no validation is performed on the data
and is sent as text/plain content type.