Add verbose detailed logging functionality
Added the option to display detailed logging of the execution of
workflow config defintions. This is so that the user will be able to
find out any issues in the config.

For example, one would be able find out any issue in the config, that is
causing the http requests to fail.