Fix the relative paths in config issue
For now, the current working directory is updated to the config
directory, when the config is parsed. This will make all the paths in
the config file relative to the config directory, rather than the
directory where the command is run.

The current working directory is resetted back to old working directory
after the parsing is completed.