Make output optional and selectable
In config, the output field can be used to define variables that needs
to be included in the output. This will only output the defined
variables. Expressions can be used as the value of the variable when
defining.

To output all the variables defined and updated during execution,
output_all field needs to be set to true. By default it is set to false.

If both of these are not defined no output is send to stdout. If output
flag is used and piped config expect a variables from the last executed
config, it might result in error.